Change database add user update account

+++ src/main/java/com/example/banking/Repository/BillRepository.java
@@ -3,5 +3,5 @@ package com.example.banking.Repository;
 import com.example.banking.Entity.BillClass;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BillRepository extends JpaRepository<BillClass, Integer> {
+public interface BillRepository extends JpaRepository<BillClass, Long> {
 }
